# loan_deal_app/src/main/java/ru/nkashlev/loan_deal_app/deal/controllers/LoanOfferController.java
package ru.nkashlev.loan_deal_app.deal.controllers;

import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;
import ru.nkashlev.loan_deal_app.deal.api.OfferApi;
import ru.nkashlev.loan_deal_app.deal.exceptions.ResourceNotFoundException;
import ru.nkashlev.loan_deal_app.deal.model.LoanOfferDTO;
import ru.nkashlev.loan_deal_app.deal.service.OfferService;

@RestController
@RequiredArgsConstructor
public class LoanOfferController implements OfferApi {
    private final OfferService offerService;

    @Override
    public ResponseEntity<Void> loanOffer(@RequestBody LoanOfferDTO request) {
        try {
            offerService.updateApplication(request);
            return ResponseEntity.ok().build();
        } catch (ResourceNotFoundException e) {
            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null); // return 404, with null body
        }
    }
}